+++ CHANGES.md
@@ -1,6 +1,8 @@
 unreleased
 -------------------
 
+- raising an exception does no longer cancel the whole context free phase(#453, @burnleydev1)
+
 - Sort embedded errors that are appended to the AST by location so the compiler
   reports the one closer to the beginning of the file first. (#463, @NathanReb)